+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 1.3.0 / ####-##-##
 ==================
 
+  * WaitForRequest and waitForResponse will resolve if the request was already made.
+  * WaitForNextRequest and waitForNextResponse added with the past behavior of waitForRequest/Response
+
 1.2.0 / 2018-09-29
 ==================